requirements.txt

django-taggit==1.1.0
django-bootstrap4==1.0.1
django-crispy-forms==1.8.1